1) Updated copyright stuff... 2) Removed Parameter classes from MicroORM and ORM code. Instead added various Parameter types (equal, not equal, and, or, like, between, etc.) under SQL namespace that both bits of code now use (also usable with SQLHelper now, but generally would only be used by MicroORM and ORM).
Note that #2 will break some code, but just switch your Parameter<Whatever> to EqualParameter<Whatever> and you have the same thing (also I added parameter types specifically for strings so for strings, you'll need to use StringEqualParameter<Whatever>). But this change adds a lot more depth to the querying power of the MicroORM and ORM and lets me expand this in the future as needed.