1) Switched DbDataReader extensions to IDataReader.
1) Switched DbDataReader extensions to IDataReader.
2) Modified Database code. It now holds a bit more info and can pull
connection string info from the config file.
3) Started to modify Mapping class to simply hold info. Most of the
functions are going to be moved to SQLHelper.
4) Started modifying SQLHelper to be simpler and faster. Also items from
the MicroORM are being moved over.

1) Updated SQLHelper some more (working towards making batching of
commands easier). A lot of the data is being moved to sub classes.
2) Added ParameterBase class
3) Created StringParameter and Parameter classes.
4) Moved the command class while I figure out how SQLHelper will work
internally.
5) Added BatchID property to IParameter (which will be used when
batching of commands).

1) Changed DbCommand's BeginTransaction to actually store the
transaction in the command...
2) Lots of modifications of SQLHelper, Command, etc. to add batching...
Still not done though.

1) Modified a number of items to get both the batching and MicroORM
items over to SQLHelper (still a number of unit tests need to be
rewritten to work with the surface changes of the class).

1) Finished updates to SQLHelper such that batching and MicroORM items
are now a part of it.