Fixed issues with saving in the ORM
Changed how conversion code works, it should now parse Enums properly.
Changed DataMapper for IDictionary<string,object> items (basically
ExpandoObject and Dynamo). It no longer copies null objects from the
dictionary.
Fixed hashing code for Dynamo so that it doesn't check for overflow.
Fixed Is extension if the base type is an object, it now returns true.
Added function to propertybase that returns the parameter version of the
value (ID value for classes, otherwise the value of the property).
Fixed some schema generation issues and added the ability to state the
scale of a decimal column (precision is determined from scale as scale*2
with a min of 18 and max of 38).