Started working on API builder
I guess this would be an OAM? Object to API mapper?
Added DeserializationFilter action filter.
Added some fixes for serialization in controller base class
Fixed GeneralValueProvider so that it doesn't leave the request stream
at the end (screws up some other items).
Added Any function for QueryProvider that can take an ID and load that
item.
Removed extension methods from Dynamo. It became difficult to serialize
the data type with the extension methods.
Fixed CompilerBase so that it loads types properly if it doesn't need to
regenerate the DLL.
Added Copy function to Dynamo. Basically it lets the Dynamo object copy
properties from another object.
Fixed ToString for Dynamo if a property is null.
Fixed issues with Left and Right extension methods for string if the
length was below 0.