Couple of speed increases
Fixed an issue with ObjectBaseClass when using the Random.NextClass
function.
Added the ability to specify a min/max range for DateTimeGenerator.
Added some checks for the mapping manager in the ORM in case of 0
mappings found for a source.
Added some checks to the IoC loader to help speed that up a bit more.
Still slower than I would like though.